backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mysqlclient==2.2.7
pillow==11.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.4.5
urllib3==2.3.0
yarl==1.18.3
